move odemis.gui.util.raster to odemis.util.raster
In fa693f0431ed23cc2e39418fbc621693e41945c1 an import of odemis.gui.util.raster was added to _tiledacq to be able to use point_in_polygon there.
Normally the .acq should not have any import from .gui. It is useful to have point_in_polygon in _tiledacq, therefore this commit moves gui.util.raster to util.raster.
