+++ src/odemis/util/raster.py
